+++ vite.config.js
@@ -1,6 +1,20 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react-swc';
 
+// Custom plugin to preserve JSDoc comments
+const preserveCommentsPlugin = () => ({
+	name: 'preserve-comments',
+	renderChunk(code) {
+		// Match and retain JSDoc comments
+		const retainedComments = code.match(/\/\*\*[\s\S]*?\*\//g) || [];
+		const preservedCode = retainedComments.join('\n') + '\n' + code;
+		return {
+			code: preservedCode,
+			map: null,
+		};
+	},
+});
+
 // https://vitejs.dev/config/
 export default defineConfig({
 	plugins: [
@@ -13,7 +27,7 @@ export default defineConfig({
 		lib: {
 			entry: 'src/index.js',
 			name: 'nanowebReact', // The global variable name for your library if used via <script> tag
-			fileName: (format) => `nanoweb-react.${format}.js`
+			fileName: (format) => `nanoweb-react.${format}.js`,
 		},
 		rollupOptions: {
 			// Ensure external dependencies like React are not bundled into your library
@@ -21,9 +35,11 @@ export default defineConfig({
 			output: {
 				globals: {
 					react: 'React',
-					'react-dom': 'ReactDOM'
-				}
-			}
-		}
-	}
+					'react-dom': 'ReactDOM',
+				},
+			},
+			plugins: [preserveCommentsPlugin()],
+		},
+		// sourcemap: true, // Enable source maps for debugging
+	},
 });